refactor(client-side-cache): move ClientSideCache interface to API module
- Move ClientSideCache interface from cocache-core to cocache-api module
- Rename interface to ComputedClientSideCache in cocache-core
- Update related classes and tests to use new interface
- Maintain backward compatibility by keeping ClientSideCache in API module
