+++ cocache-core/src/main/kotlin/me/ahoo/cache/CacheManager.kt
@@ -13,8 +13,8 @@
 
 package me.ahoo.cache
 
+import me.ahoo.cache.api.client.ClientSideCache
 import me.ahoo.cache.api.source.CacheSource
-import me.ahoo.cache.client.ClientSideCache
 import me.ahoo.cache.client.MapClientSideCache
 import me.ahoo.cache.consistency.CacheEvictedEventBus
 import me.ahoo.cache.converter.KeyConverter
