+++ cocache-core/src/test/kotlin/me/ahoo/cache/CoherentCacheTest.kt
@@ -12,7 +12,7 @@
  */
 package me.ahoo.cache
 
-import me.ahoo.cache.client.ClientSideCache
+import me.ahoo.cache.api.client.ClientSideCache
 import me.ahoo.cache.client.MapClientSideCache
 import me.ahoo.cache.consistency.CacheEvictedEventBus
 import me.ahoo.cache.consistency.GuavaCacheEvictedEventBus
