+++ cocache-example/src/main/kotlin/me/ahoo/cache/example/config/UserCacheConfiguration.kt
@@ -13,7 +13,7 @@
 package me.ahoo.cache.example.config
 
 import me.ahoo.cache.api.source.CacheSource
-import me.ahoo.cache.client.ClientSideCache
+import me.ahoo.cache.api.client.ClientSideCache
 import me.ahoo.cache.client.MapClientSideCache
 import me.ahoo.cache.example.model.User
 import org.springframework.context.annotation.Bean
