+++ cocache-spring-boot-starter/src/test/kotlin/me/ahoo/cache/spring/boot/starter/customize/EnableCoCacheConfigurationWithCustomize.kt
@@ -15,7 +15,7 @@ package me.ahoo.cache.spring.boot.starter.customize
 
 import io.mockk.mockk
 import me.ahoo.cache.api.source.CacheSource
-import me.ahoo.cache.client.ClientSideCache
+import me.ahoo.cache.client.ComputedClientSideCache
 import me.ahoo.cache.example.cache.UserCache
 import me.ahoo.cache.example.model.User
 import me.ahoo.cache.spring.EnableCoCache
@@ -31,7 +31,7 @@ import org.springframework.context.annotation.Bean
 class EnableCoCacheConfigurationWithCustomize {
 
     @Bean
-    fun customizeUserClientSideCache(): ClientSideCache<User> {
+    fun customizeUserClientSideCache(): ComputedClientSideCache<User> {
         return mockk()
     }
