+++ cocache-test/src/main/kotlin/me/ahoo/cache/test/MultipleInstanceSyncSpec.kt
@@ -16,7 +16,7 @@ package me.ahoo.cache.test
 import me.ahoo.cache.CacheConfig
 import me.ahoo.cache.CacheManager
 import me.ahoo.cache.CoherentCache
-import me.ahoo.cache.client.ClientSideCache
+import me.ahoo.cache.api.client.ClientSideCache
 import me.ahoo.cache.consistency.CacheEvictedEvent
 import me.ahoo.cache.consistency.CacheEvictedEventBus
 import me.ahoo.cache.consistency.CacheEvictedSubscriber
